% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_factors.R
\name{detect_factors}
\alias{detect_factors}
\title{find if a factor is actually character}
\usage{
detect_factors(f)
}
\arguments{
\item{x}{a factor}
}
\value{
boolean; \code{TRUE} if input is factor,
        \code{FALSE} if input is character
}
\description{
In order to check if a factor is a character,this function checks
}
\details{
whether there are repeated value in the factor
}
\examples{
detect_factors(factor(c("a","b","a"))) # True
detect_factors(factor(c("a","b","c"))) # False
}
